test_model.metrics: compute the mean over elements instead of the sum
Jess told us the *sum* of linex loss was computed over all ements but
after the third submission it seems very likely that the *mean* error
over all elements is computed instead.

I reflect this change in our way to compute the linex loss
in linex_error, returning the mean instead of the sum.