tools.create_submission: offer the possibility not to provide x_test
1 - For example,

>>> import tools
>>> import numpy as np
>>> y_predict = np.ones((82909, 1))
>>> tools.create_submission(y_predict)

Will create a submission file in submissions/ folder, with a prediction
of 1 received calls for every row of the submission file.

2 - Versioning now ignore files created by create_submission function
Only commit submission files that has been submitted to the leaderboard
to the repository. What I suggest is to give an incremental id to the
submission renaming it manually, for example :

submission_20161123_145909.txt -> submission4.txt

Then do not forget to add a log in the google drive submission_log tab
for this new submission, and commit it.